+++ app/controllers/api/v1/characters_controller.rb
@@ -129,7 +129,6 @@ module Api
 
       def create_additional_structures_for_character(character)
         CreateCharacterRoles.call(character_id: character.id, character_role_params: character_role_params)
-        CreateDungeonAccess.call(character_id: character.id, dungeon_params: dungeon_params)
         CreateCharacterProfessions.call(character_id: character.id, profession_params: profession_params)
         CreateGuildInvite.call(character: character, guild: Guild.find_by(world_fraction: character.world_fraction, id: params[:character][:guild_id]), from_guild: false) if params[:character][:guild_id].present?
       end
@@ -143,10 +142,6 @@ module Api
         h[:guild] = @character.nil? ? nil : @character&.guild
         h[:user] = Current.user
         h
-      end
-
-      def dungeon_params
-        params.require(:character).permit(dungeon: {})[:dungeon]
       end
 
       def character_role_params