+++ app/models/dungeon_access.rb
@@ -1,5 +1,0 @@
-# Represents character accesses to dungeons and raids
-class DungeonAccess < ApplicationRecord
-  belongs_to :character
-  belongs_to :dungeon
-end