+++ app/serializers/character_show_serializer.rb
@@ -1,5 +1,5 @@
 class CharacterShowSerializer < ActiveModel::Serializer
-  attributes :id, :name, :level, :character_class_id, :race_id, :guild_id, :world_id, :main_role_id, :secondary_role_ids, :dungeon_ids, :profession_ids
+  attributes :id, :name, :level, :character_class_id, :race_id, :guild_id, :world_id, :main_role_id, :secondary_role_ids, :profession_ids
 
   def main_role_id
     object.main_roles&.first&.id
@@ -7,10 +7,6 @@ class CharacterShowSerializer < ActiveModel::Serializer
 
   def secondary_role_ids
     object.secondary_roles.pluck(:id)
-  end
-
-  def dungeon_ids
-    object.dungeons.pluck(:id)
   end
 
   def profession_ids