+++ app/serializers/dungeon_serializer.rb
@@ -1,3 +1,3 @@
 class DungeonSerializer < ActiveModel::Serializer
-  attributes :id, :name, :raid, :key_access, :quest_access
+  attributes :id, :name, :raid
 end