+++ spec/factories/dungeon_accesses.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :dungeon_access do
-    association :character, :human_warrior
-    association :dungeon
-  end
-end