+++ spec/models/character_spec.rb
@@ -5,8 +5,6 @@ RSpec.describe Character, type: :model do
   it { should belong_to :world }
   it { should belong_to(:guild).optional }
   it { should belong_to :world_fraction }
-  it { should have_many(:dungeon_accesses).dependent(:destroy) }
-  it { should have_many(:dungeons).through(:dungeon_accesses) }
   it { should have_many(:owned_events).class_name('Event').with_foreign_key('owner_id').dependent(:destroy) }
   it { should have_many(:subscribes).dependent(:destroy) }
   it { should have_many(:events).through(:subscribes).source(:subscribeable) }